Split `input`/`inputFile` options to their own file
