+++ lib/stdio/async.js
@@ -1,6 +1,6 @@
 import {createReadStream, createWriteStream} from 'node:fs';
 import {Readable, Writable} from 'node:stream';
-import {handleInput} from './input.js';
+import {handleInput} from './handle.js';
 
 // Handle `input`, `inputFile`, `stdin`, `stdout` and `stderr` options, before spawning, in async mode
 export const handleInputAsync = options => handleInput(addPropertiesAsync, options);
