+++ lib/stdio/sync.js
@@ -1,5 +1,5 @@
 import {readFileSync, writeFileSync} from 'node:fs';
-import {handleInput} from './input.js';
+import {handleInput} from './handle.js';
 import {TYPE_TO_MESSAGE} from './type.js';
 
 // Handle `input`, `inputFile`, `stdin`, `stdout` and `stderr` options, before spawning, in sync mode
